# app/swagger.yaml
openapi: 3.0.0
info:
  title: API Documentation
  version: 1.0.0
paths:
...
  /resumes:
    post:
      summary: Create a resume
      description: Creates a new resume for the authenticated user.
      tags:
        - Resumes
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                resume_link:
                  type: string
                  description: The link to the resume file.
                  example: "https://example.com/resume.pdf"
      responses:
        201:
          description: Resume created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Resume created successfully
                  resume_id:
                    type: integer
                    example: 1
        400:
          description: Missing required fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Resume Link is required
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Internal server error

    get:
      summary: Get all resumes
      description: Retrieves all resumes for the authenticated user with pagination support.
      tags:
        - Resumes
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
          description: The page number for pagination.
        - name: per_page
          in: query
          required: false
          schema:
            type: integer
            default: 20
          description: The number of resumes per page.
      responses:
        200:
          description: List of resumes
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        resume_link:
                          type: string
                          example: "https://example.com/resume.pdf"
                        created_at:
                          type: string
                          format: date-time
                          example: "2024-12-03T12:34:56"
                  pagination:
                    type: object
                    properties:
                      current_page:
                        type: integer
                        example: 1
                      per_page:
                        type: integer
                        example: 20
                      total_items:
                        type: integer
                        example: 3
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Internal server error
  /resumes/{resume_id}:
    put:
      summary: Update a resume
      description: Updates the resume link of a specific resume for the authenticated user.
      tags:
        - Resumes
      security:
        - BearerAuth: []
      parameters:
        - name: resume_id
          in: path
          required: true
          description: The ID of the resume to update.
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                resume_link:
                  type: string
                  description: The new link to the resume file.
                  example: "https://example.com/updated-resume.pdf"
      responses:
        200:
          description: Resume updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Resume updated successfully
        400:
          description: No fields to update provided
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: No fields to update provided
        404:
          description: Resume not found or unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Resume not found or unauthorized
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Internal server error
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT